+++ src/main/java/io/github/torand/fastersql/projection/Projections.java
@@ -1,9 +1,16 @@
 package io.github.torand.fastersql.projection;
 
+import io.github.torand.fastersql.statement.SelectStatement;
+import io.github.torand.fastersql.subquery.ExpressionSubquery;
+
 public final class Projections {
     private Projections() {}
 
     public static ColumnPosition colPos(int position) {
         return new ColumnPosition(position);
+    }
+
+    public static ExpressionSubquery subquery(SelectStatement query) {
+        return new ExpressionSubquery(query);
     }
 }