+++ src/main/java/io/github/torand/fastersql/subquery/Subquery.java
@@ -19,58 +19,35 @@ import io.github.torand.fastersql.Column;
 import io.github.torand.fastersql.Context;
 import io.github.torand.fastersql.Sql;
 import io.github.torand.fastersql.alias.ColumnAlias;
-import io.github.torand.fastersql.alias.TableAlias;
 import io.github.torand.fastersql.statement.SelectStatement;
 
 import java.util.stream.Stream;
 
 import static io.github.torand.fastersql.util.collection.CollectionHelper.streamSafely;
-import static io.github.torand.fastersql.util.contract.Requires.requireNonBlank;
-import static java.util.Objects.requireNonNull;
 
-public class Subquery implements Sql {
-    private final SelectStatement selectStatement;
-    private final TableAlias alias; // When used as projection (scalar value) this is interpreted as a column alias
+public interface Subquery extends Sql {
 
-    public Subquery(SelectStatement selectStatement) {
-        this.selectStatement = requireNonNull(selectStatement, "No select statement specified");
-        this.alias = null;
-    }
-
-    private Subquery(SelectStatement selectStatement, String alias) {
-        this.selectStatement = requireNonNull(selectStatement, "No select statement specified");
-        this.alias = new TableAlias(requireNonBlank(alias, "No alias specified"));
-    }
+    SelectStatement selectStatement();
 
     // Sql
 
     @Override
-    public String sql(Context context) {
-        return "(" + selectStatement.sql(context) + ")";
+    default String sql(Context context) {
+        return "(" + selectStatement().sql(context) + ")";
     }
 
     @Override
-    public Stream<Object> params(Context context) {
-        return streamSafely(selectStatement.params(context));
+    default Stream<Object> params(Context context) {
+        return streamSafely(selectStatement().params(context));
     }
 
     @Override
-    public Stream<Column> columnRefs() {
+    default Stream<Column> columnRefs() {
         return Stream.empty();
     }
 
     @Override
-    public Stream<ColumnAlias> aliasRefs() {
+    default Stream<ColumnAlias> aliasRefs() {
         return Stream.empty();
-    }
-
-    // "Projection"
-
-    public Subquery as(String alias) {
-        return new Subquery(selectStatement, alias);
-    }
-
-    public String alias() {
-        return alias.name();
     }
 }